updated rust template
